feat: move sable util fns to network folder
